isa/evaluation/requirements.txt
PyYAML
google-cloud-aiplatform